XRDDEV-843 Unregister auth cert or mark for deletion // PR fix 1
* better exception handling with ManagementRequestSendingFailedException
