+++ src/proxy-ui-api/src/main/java/org/niis/xroad/restapi/service/ManagementRequestSenderService.java
@@ -31,6 +31,7 @@ import ee.ria.xroad.common.identifier.SecurityServerId;
 import ee.ria.xroad.common.request.ManagementRequestSender;
 
 import lombok.extern.slf4j.Slf4j;
+import org.niis.xroad.restapi.exceptions.ErrorDeviation;
 import org.niis.xroad.restapi.facade.GlobalConfFacade;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -86,16 +87,13 @@ public class ManagementRequestSenderService {
      * @param authCert the authentication certificate bytes
      * @return request ID in the central server database (e.g. for audit logs if wanted)
      */
-    public Integer sendAuthCertDeletionRequest(SecurityServerId securityServer, byte[] authCert)
-            throws GlobalConfService.GlobalConfOutdatedException {
+    public Integer sendAuthCertDeletionRequest(SecurityServerId securityServer, byte[] authCert) throws
+            GlobalConfService.GlobalConfOutdatedException, ManagementRequestSendingFailedException {
         ManagementRequestSender sender = createManagementRequestSender();
         try {
             return sender.sendAuthCertDeletionRequest(securityServer, authCert);
         } catch (Exception e) {
-            if (e instanceof CodedException) {
-                throw (CodedException) e;
-            }
-            throw new RuntimeException(e);
+            throw new ManagementRequestSendingFailedException(e);
         }
     }
 
@@ -106,5 +104,20 @@ public class ManagementRequestSenderService {
         ClientId sender = serverConf.getOwner().getIdentifier();
         ClientId receiver = globalConfFacade.getManagementRequestService();
         return new ManagementRequestSender(sender, receiver);
+    }
+
+    /**
+     * Missing a valid auth cert
+     */
+    public static class ManagementRequestSendingFailedException extends ServiceException {
+        public static final String MANAGEMENT_REQUEST_SENDING_FAILED = "management_request_sending_failed";
+
+        public ManagementRequestSendingFailedException(Throwable t) {
+            super(t, createError(t));
+        }
+
+        private static ErrorDeviation createError(Throwable t) {
+            return new ErrorDeviation(MANAGEMENT_REQUEST_SENDING_FAILED, t.getMessage());
+        }
     }
 }
